Final Revision
Using a remote SQL server connection.

This revision checks for prerequisites, timeSlotID conflicts, duplicate courses, and open seats.

UI changes made for a better user experience (Event handlers refreshing the list boxes)

I had to make a new fork for some reason because mine is not considered a fork of the original project for some reason...